Conductores version1 terminada

+++ CI 18-2 MiniProy2 - Inicial/src/main/java/co/edu/icesi/mio/logic/TmioConductoresLogic.java
@@ -2,10 +2,12 @@ package co.edu.icesi.mio.logic;
 
 import java.util.ArrayList;
 import java.util.Calendar;
+import java.util.Date;
 import java.util.GregorianCalendar;
 import java.util.List;
 
 import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -22,37 +24,100 @@ public class TmioConductoresLogic implements ITmioConductoresLogic{
 	@Autowired
 	private ITmio1_Conductores_DAO conductorDAO;
 	
+	@PersistenceContext
+    private EntityManager em;
 	
 	@Transactional
-	public void createConductor(Tmio1Conductore conductor, EntityManager em) {
+	public void createConductor(Tmio1Conductore conductor) {
 		//hacer validaciones sobre conductor
 		//NO ESTOY SEGURA DE PASAR EL ENTITYMANAGER COMO PARAMETRO O DEBE CREARSE.
-		conductorDAO.save(em,conductor);
+		
+		if(conductor != null) {
+			if(validacionCedula(conductor.getCedula()) && 
+					validacionNombre(conductor.getNombre()) &&
+						validacionApellido(conductor.getApellidos()) &&
+							validacionFechaNacimiento(conductor.getFechaNacimiento()) &&
+								validacionFechaContratacion(conductor.getFechaContratacion())) {
+				conductorDAO.save(em,conductor);
+			}
+		}else {
+			//LANZAR UNA EXCEPCION
+		}	
 	}
-
+	
 	@Override
-	public void updateConductor(Tmio1Conductore conductor, EntityManager em) {
-		conductorDAO.update(em, conductor);
+	public void updateConductor(Tmio1Conductore conductor) {
+		if(conductor != null) {
+			if(validacionCedula(conductor.getCedula()) && 
+					validacionNombre(conductor.getNombre()) &&
+						validacionApellido(conductor.getApellidos()) &&
+							validacionFechaNacimiento(conductor.getFechaNacimiento()) &&
+								validacionFechaContratacion(conductor.getFechaContratacion())) {
+				conductorDAO.update(em, conductor);
+			}
+		}else {
+			//LANZAR UNA EXCEPCION
+		}
 	}
 
 	@Transactional
-	public void deleteConductore(Tmio1Conductore conductor, EntityManager em) {
+	public void deleteConductor(Tmio1Conductore conductor) {
+		if(conductor!=null)
 		conductorDAO.delete(em, conductor);
 	}
 
 	@Transactional
-	public List<Tmio1Conductore> findByName(String name, EntityManager em) {
-		return conductorDAO.findByName(em, name);
+	public List<Tmio1Conductore> findByName(String name) {
+		List<Tmio1Conductore> conductores= null;
+		if(validacionNombre(name))
+			conductores= conductorDAO.findByName(em, name);
+		return conductores;
 	}
 
 	@Transactional
-	public List<Tmio1Conductore> findByLastname(String lastname, EntityManager em) {
-		return conductorDAO.findByLastName(em, lastname);
+	public List<Tmio1Conductore> findByLastname(String lastname) {
+		List<Tmio1Conductore> conductores= null;
+		if(validacionNombre(lastname))
+			conductores= conductorDAO.findByLastName(em, lastname);
+		return conductores; 
 	}
 
 	@Transactional
-	public Tmio1Conductore findByCedula(String cedula, EntityManager em) {
-		return conductorDAO.findByCedula(em, cedula);
+	public Tmio1Conductore findByCedula(String cedula) {
+		Tmio1Conductore conductor= null;
+		if(validacionNombre(cedula))
+			conductor= conductorDAO.findByCedula(em, cedula);
+		return conductor; 
+	}
+	
+	/**
+	 * 	VALIDACIONES
+	 */
+	public boolean validacionCedula(String cedula) {
+		return cedula.matches("[0-9]+");
+	}
+	
+	public boolean validacionNombre(String nombre) {
+		return !nombre.equals("") && nombre.length()>=3;
+	}
+	
+	public boolean validacionApellido(String apellido) {
+		return !apellido.equals("") && apellido.length()>=3;
+	}
+	
+	public boolean validacionFechaNacimiento(Date fechaNacimiento) {
+		Calendar gc1= new GregorianCalendar();
+		gc1.setTime(fechaNacimiento);
+		Calendar gc= new GregorianCalendar().getInstance();
+		return fechaNacimiento!= null && (gc.get(GregorianCalendar.YEAR) - gc1.get(GregorianCalendar.YEAR))>=18;
+	}
+	
+	public boolean validacionFechaContratacion(Date fechaContratacion) {
+		Calendar gc1= new GregorianCalendar();
+		gc1.setTime(fechaContratacion);
+		Calendar gc= new GregorianCalendar().getInstance();
+		
+		return fechaContratacion != null && gc1.compareTo(gc)<0;
 	}
